lua compiler: Implement initial tokeniser
This is the first pass of the parser's tokeniser. The `FileTokeniser` class reads the entire contents of a file when it is constructed, and will then tokenise it a token at a time whenever it's `eat` method is called.

The patterns used to `eat` and `peek` most tokens are specified in Statue.lua, however END_OF_FILE, STRING_LITERAL, and NUMBER_LITERAL are all special cases.