lua compiler: Implemented Reference class
These don't do anything except override how the value they reference is serialised to json. Instead of the value being serialised, it is converted to a string and then serialised. As tables in lua can have cyclic references to each other, this is a way to prevent json serialisation recurring indefinitely when attempting to serialise a cyclic data structure.